Factorize code between Gitlab and GitHub publisher
