+++ scripts/dataset/index.js
@@ -6,7 +6,6 @@ import config from 'config';
 import generateRelease from './export/index.js';
 import logger from './logger/index.js';
 import publishRelease from './publish/index.js';
-import publishReleaseGitLab from './publishGitLab/index.js';
 
 export async function release({ shouldPublish, shouldRemoveLocalCopy, fileName }) {
   const releaseDate = new Date();
@@ -15,15 +14,7 @@ export async function release({ shouldPublish, shouldRemoveLocalCopy, fileName }
 
   logger.info('Start exporting dataset…');
 
-  const usesGitHub = (typeof process.env.OTA_ENGINE_GITHUB_TOKEN !== 'undefined');
-  const usesGitLab = (typeof process.env.OTA_ENGINE_GITLAB_TOKEN !== 'undefined');
-
-  let versionsRepositoryURL = '';
-
-  if (usesGitHub) versionsRepositoryURL = config.get('@opentermsarchive/engine.dataset.versionsRepositoryURL');
-  if (usesGitLab) versionsRepositoryURL = config.get('@opentermsarchive/engine.dataset.versionsRepositoryURLGitLab');
-
-  const stats = await generateRelease({ archivePath, releaseDate, versionsRepositoryURL });
+  const stats = await generateRelease({ archivePath, releaseDate });
 
   logger.info(`Dataset exported in ${archivePath}`);
 
@@ -33,25 +24,13 @@ export async function release({ shouldPublish, shouldRemoveLocalCopy, fileName }
 
   logger.info('Start publishing dataset…');
 
-  if (usesGitHub) {
-    const releaseUrl = await publishRelease({
-      archivePath,
-      releaseDate,
-      stats,
-    });
+  const releaseUrl = await publishRelease({
+    archivePath,
+    releaseDate,
+    stats,
+  });
 
-    logger.info(`Dataset published to ${releaseUrl}`);
-  }
-
-  if (usesGitLab) {
-    const releaseUrl = await publishReleaseGitLab({
-      archivePath,
-      releaseDate,
-      stats,
-    });
-
-    logger.info(`Dataset published to ${releaseUrl}`);
-  }
+  logger.info(`Dataset published to ${releaseUrl}`);
 
   if (!shouldRemoveLocalCopy) {
     return;
